+++ CHANGES.md
@@ -5,6 +5,7 @@ Here you can see the full list of changes between each Pykit release.
 
 Version 0.3.8
 -----------
+- added: select multiple files with html5
 - changed: uploader settings moved to `uploadOptions` config
 
 Version 0.3.7