+++ CHANGES.md
@@ -6,6 +6,7 @@ Here you can see the full list of changes between each Pykit release.
 Version 0.3.8
 -----------
 - added: select multiple files with html5
+- added: `position` config now supports `absolute`
 - changed: uploader settings moved to `uploadOptions` config
 
 Version 0.3.7